added topics model

+++ db/migrate/20190506150353_devise_token_auth_create_users.rb
@@ -34,7 +34,7 @@ class DeviseTokenAuthCreateUsers < ActiveRecord::Migration[5.2]
       t.string :first_name
       t.string :last_name
       t.string :username
-      t.numeric :gender
+      t.integer :gender
 
       ## Tokens
       t.json :tokens
